ai_rag_simple/requirements.txt
torch
transformers
faiss-cpu
numpy
datasets
sentence-transformers

